+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - SIMPLY-4066: Install Reservoir (NYPL Design System) and use components to build UI.
 - SIMPLY-4126: Fix bug where "Registry Stage" dropdown was showing "testing" when it should have shown "canceled."
 - SIMPLY-4125: Add patron counts to simple list view.
+- SIMPLY-4172: Add library description to library details.
 
 ## 1.4.15